Add submap registration constraints in both directions
Submap registration constraints from A->B and B->A are not symmetric, since the ESDF is only propagated into known space. This means it is usually propagated up to the max_esdf_distance (e.g. 2m) from the surface in the camera direction but only up to the TSDF truncation distance (e.g. 0.6m) behind the surface.
This can cause issues when the isoruface points from a submap A submap fall more than 0.6m behind the isosurface of another submap B.
A simple fix is to add constraints from both A->B and B->A, since the isosurface points of B would still fall in front of the surface of A.
Although this doubles the evaluation cost for constraints of submap pair (A, B), it seems to help the optimization to converge in less steps. Therefore the increase in total computational cost seems acceptable. On platforms where the computing power is very limited, one could try using a sampling ratio that is even lower than the default 5% (e.g. 2.5%).